Remove old interpreter manager
