--- src/vs/workbench/browser/parts/positronTopActionBar/components/topActionBarInterpretersManager.css
@@ -1,70 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2022 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-.top-action-bar-interpreters-manager {
-	flex: 0;
-	width: min-content;
-	margin: 0 6px;
-	height: 26px;
-	display: grid;
-	cursor: pointer;
-	align-items: center;
-	border-radius: 6px;
-	box-sizing: border-box;
-	grid-template-columns: 1fr 1fr;
-	color: var(--positronActionBar-foreground);
-	border: 1px solid var(--vscode-positronTopActionBar-selectBorder);
-	background-color: var(--vscode-positronTopActionBar-selectBoxBackground);
-}
-
-.top-action-bar-interpreters-manager:hover {
-	background-color: var(--positronActionBar-hoverBackground);
-}
-
-.top-action-bar-interpreters-manager .left {
-	margin-left: 1px;
-}
-
-.top-action-bar-interpreters-manager .left .label {
-	margin: 0 4px;
-	padding: 0 3px;
-	height: 22px;
-	border: none;
-	display: flex;
-	font-size: 12px;
-	cursor: pointer;
-	overflow: visible;
-	white-space: nowrap;
-	align-items: center;
-	background: transparent;
-	justify-content: center;
-	color: var(--positronActionBar-foreground);
-}
-
-.top-action-bar-interpreters-manager .left .label .icon {
-	width: 16px;
-	height: 16px;
-	margin-right: 6px;
-}
-
-.top-action-bar-interpreters-manager .right {
-	display: flex;
-	align-items: center;
-	justify-content: right;
-	margin-right: 1px;
-}
-
-.top-action-bar-interpreters-manager .right .chevron {
-	height: 22px;
-	width: 22px;
-	font-size: 14px;
-	border: none;
-	display: flex;
-	cursor: pointer;
-	overflow: visible;
-	align-items: center;
-	background: transparent;
-	justify-content: center;
-}
